requirements.txt
crewai
load_dotenv
langchain_google_genai
crewai_tools
streamlit>=1.32.0
plotly>=5.18.0
streamlit-lottie>=0.0.5
tavily-python